Charting System ready to go
This was really hard to make (even with the copying stuff) and it would take a long time to finish, time that I unfortunately do not have. So i took ninjaMuffin code (i’m bad at coding lmao). Of course, it was re-adapted to fit with the gameplay. In the start of every hx file I copied, I put a little reference to the original author.

This is Just a prototype, if it turns into a finished project, a lot of code will need to be changed (plagiarism is for noobs). Anyways, thx for the help ninjaMuffin, i wouldn’t be able to finish this in time (i have about 2 days to end this, pls help) (also, i need to learn to code more complicated stuff, with real Math involved, will add a final thoughts note at the end of this grind)